+++ appium/sample-scripts/java/src/test/java/com/testdroid/appium/android/sample/AndroidAppiumExampleTest.java
@@ -17,11 +17,9 @@ public class AndroidAppiumExampleTest  extends BaseAndroidTest {
         setUpTest();
     }
     @AfterClass
-    public static void tearDown()
+    public void tearDown()
     {
-        if (wd != null) {
-            wd.quit();
-        }
+        quitAppiumSession();
     }
